Deleting Wish List tests added

+++ functions/src/wishes/wish.service.ts
@@ -11,8 +11,8 @@ export class WishService {
     if(uid === null)
       return Promise.reject('User uid can NOT be null when creating wishlist');
 
-    let exist = await this.repo.wishWithIdExists(uid);
-    
+    let exist = await this.repo.hasWishlistWithId(uid);
+
     if(exist === true)
       return Promise.reject('This user already has a wishlist created');
 
@@ -20,6 +20,13 @@ export class WishService {
   }
 
   deleteUserWishes(uid: string): Promise<any> {
+
+    if(uid === undefined)
+      return Promise.reject('User uid can NOT be undefined on delete wishlist request');
+
+    if(uid === null)
+      return Promise.reject('User uid can NOT be null on delete wishlist request');
+
     return this.repo.deleteUserWishes(uid);
   }
 }
